feat: Added Player domain object with related enums.
For creation of a player it is necessary to specify the available game play options and an optional id. The player is storing an internal game result store which is updated by the recordGame method. The current result store state can be returned.

Added functionality for evaluate game result for 2 play options.

The method for evaluation is placed to the PlayOption enum to be able
to get the result for en existing PlayOption instead of putting 2 PlayOptions to a function.

The Player is implemented as class (with custom equals - and hashCode) because data classes are leaking the copy- constructor and it should not be possible to create invalid instances.